grammar/requirements.txt
flask==3.1.0
werkzeug==3.1.3
flask-cors==5.0.1
torch==2.6.0
transformers==4.49.0
openai-whisper==20240930
pydub==0.25.1
psutil==5.9.6
gunicorn==23.0.0
ffmpeg-python==0.2.0
numpy==1.26.4
numba==0.59.0